trivial refacor: seperate LayeredMap::new_layer() code out
